locking fixes

+++ src/handle_copymove.rs
@@ -109,9 +109,7 @@ impl super::DavHandler {
     pub(crate) fn do_move(&self, source: &WebPath, dest: &WebPath, existed: bool, mut multierror: MultiError) -> DavResult<()> {
         debug!("do_move {} {}", source, dest);
         if let Err(e) = self.fs.rename(source, dest) {
-            // XXX FIXME probably need to check if the failure was
-            // source or destionation related and produce the
-            // correct error & path.
+            // This is a single action ATM, so do not need multi-error.
             add_status(&mut multierror, source, e);
             Err(DavError::Status(multierror.close()?))
         } else {
@@ -133,25 +131,55 @@ impl super::DavHandler {
         // decode and validate destination.
         let dest = req.headers.get::<headers::Destination>()
                     .ok_or(statuserror(&mut res, SC::BadRequest))?;
-        let dest = match WebPath::from_str(&dest.0, &self.prefix) {
+        let mut dest = match WebPath::from_str(&dest.0, &self.prefix) {
             Err(e) => Err(daverror(&mut res, e)),
             Ok(d) => Ok(d),
         }?;
 
-        // source must exist, as well as the parent of the destination.
-        let path = self.path(&req);
-        let meta = self.fs.metadata(&path).map_err(|e| fserror(&mut res, e))?;
+        // get the source. for MOVE, tread with care- if the path ends in "/" but
+        // it actually is a symlink, we want to move the symlink, not what it points to.
+        let mut path = self.path(&req);
+        let meta = if method == Method::Copy {
+            self.fs.metadata(&path).map_err(|e| fserror(&mut res, e))?
+        } else {
+            path.remove_slash();
+            let meta = self.fs.symlink_metadata(&path).map_err(|e| fserror(&mut res, e))?;
+            if meta.is_dir() {
+                path.add_slash();
+            }
+            meta
+        };
+
+        // parent of the destination must exist.
         if !self.has_parent(&dest) {
             Err(statuserror(&mut res, SC::Conflict))?;
         }
 
-        // check if overwrite is "F"
-        let dmeta = if depth == Depth::Zero {
-            self.fs.metadata(&dest)
-        } else {
-            self.fs.symlink_metadata(&dest)
+        // for the destination, also check if it's a symlink. If we are going
+        // to remove it first, we want to remove the link, not what it points to.
+        let mut d = dest.clone();
+        d.remove_slash();
+        let (dest_is_file, dmeta) = match self.fs.symlink_metadata(&d) {
+            Ok(meta) => {
+                let mut is_file = false;
+                if meta.is_symlink() {
+                    if let Ok(m) = self.fs.metadata(&d) {
+                        is_file = m.is_file();
+                    }
+                }
+                if meta.is_file() {
+                    is_file = true;
+                }
+                if meta.is_dir() {
+                    d.add_slash();
+                }
+                dest = d;
+                (is_file, Ok(meta))
+            },
+            Err(e) => (false, Err(e)),
         };
 
+        // check if overwrite is "F"
         let exists = dmeta.is_ok();
         if !overwrite && exists {
             Err(statuserror(&mut res, SC::PreconditionFailed))?;
@@ -168,33 +196,34 @@ impl super::DavHandler {
             Err(s) => return Err(statuserror(&mut res, s)),
         };
 
-        // check locks XXX FIXME multistatus errors
+        let mut multierror = MultiError::new(res, &path);
+
+        // check locks. since we cancel the entire operation if there is
+        // a conflicting lock, we do not return a 207 multistatus, but
+        // just a simple status.
         if let Some(ref locksystem) = self.ls {
             let t = tokens.iter().map(|s| s.as_str()).collect::<Vec<&str>>();
             if method == Method::Move {
                 // for MOVE check if source path is locked
                 if let Err(_l) = locksystem.check(&path, true, t.clone()) {
-                    return Err(statuserror(&mut res, SC::Locked));
+                    return multierror.finalstatus(&path, SC::Locked);
                 }
             }
             // for MOVE and COPY check if destination is locked
             if let Err(_l) = locksystem.check(&dest, true, t) {
-                return Err(statuserror(&mut res, SC::Locked));
+                return multierror.finalstatus(&path, SC::Locked);
             }
         }
 
-        let mut multierror = MultiError::new(res, &path);
-
         // see if we need to delete the destination first.
-        if overwrite && exists && depth != Depth::Zero {
+        if overwrite && exists && depth != Depth::Zero && !dest_is_file {
             debug!("handle_copymove: deleting destination {}", dest);
             if let Err(_) = self.delete_items(&mut multierror, Depth::Infinity, dmeta.unwrap(), &dest) {
                 return Err(DavError::Status(multierror.close()?));
             }
-            // XXX FIXME should really do this per item, in case the
-            // delete partially fails.
+            // should really do this per item, in case the delete partially fails. See TODO.md
             if let Some(ref locksystem) = self.ls {
-                locksystem.delete(&path).ok();
+                locksystem.delete(&dest).ok();
             }
         }
